Don't force user to scan list across and down in same context [REF]
Within a function or constructor call, if it can't fit in the width
limit on one line, list downwards - don't mix the two.
Having a mixture of vertical and horizontal listing makes it very easy
to miss extra parameters that are off to the right.
Also typesetters have set text in narrow columns for hundreds of years for
a reason.