Add an account_id - the first one listed on the budgetary position
[O2549]

To allow us to group by that account id and attributes of it.

This is to work around the fact you can't group by a *2many because
they don't produce a field on self in the database.

The solution of using the first account listed on the budgetary position
is Narender's:

 Re: [O2549] Budget management changes

Nick can we use the account from first line from the account_ids on
"account.budget.post". If required, please create an invisible Many2One
of "account.account" on "account.budget.post" (whose value will be the
from the first line of account_ids )