Introduce concept of a 'main account' on a budgetary position [REF]
I believe this is more semantic in terms of what we mean than
"first_account_id".  Other objects should be able to lean on
main_account_id and we should be able to decide that it's picked using
different criteria (or even manually by the user) at a later date.
We couldn't reuse 'first_account_id' field in that way.