Revert "Switch to using proper stored computed fields rather than .write() hack"
This reverts commit a1650fdd0c46ddf28e97b20e4d2e89a431602cc9.

It looks like with proper computed fields the actual amount was not
recomputed when journal items were posted.  However, it's not clear how
it even works with the write() hack or how it makes a difference.  No
comments or notes in the commit message were provided in the original,
which is why I changed it!