[O2576] Replace auto-recompute with manual trigger (more reliable)
As it turns out, ACs implementation of the compute_practical_amount
in read_group() was not working fully as expected. Figuring out the
issues in this implementation and re-working it would have been
troublesome at best, so I've opted for a manual approach with the
approval from NG to do so.

This was it's going to be much more
performance focused (which is going to start becoming an issue in the
long term) and we are relying less on silly things like Odoo deciding
to not recompute the column because it's not applicable to the view
it's on.