[O2576] Update the client view when re-calculate button is hit
This works in all three views the button is on (list, graph, pivot)
because each one of them house a search box. I spent a while trying
to understand what is scoped in as `this` in the button click event
but nothing really useful for doing this action is there, hence
the slightly 'hacky' approach of just simulating the enter keypress
on the search box.
Note I also tried returning `{'type': 'ir.actions.client', 'tag':
'reload'} from the python func but this did not propagate up in our
custom handler and nothing happened.