[FIX] Data in uk_accounting module conflicts with that in l10n_uk
This fix works by overriding the account module's post-init-hook method at a python level to install uk_accounting module instead of l10n_uk. This will only work if the uk_accounting module files are present when the container starts. If you already have a database which is configured as UK and have the account module already installed, you will not be able to then install the uk_accounting module.