[FIX] Module cannot be used with l10n_uk
The original design of this module meant that it could not be used at the same time as l10n_uk. Because the 'account' module automatically installs l10n_uk, hacks were required to make it work. This commit refactors the whole module to inherit l10n_uk and update existing records instead of conflicting with them.